requirements.txt
Django==1.7
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
wsgiref==0.1.2
django-admin-bootstrapped==2.0.5
django-tastypie==0.12.1
python-dateutil>=2.1
lxml==3.1.0
